Various smaller changes to fix compilation and other problems.
- Added new rule to test grammar to get code generation for wildcard capture.
- Updated the Cpp.stg template file for that.
- Made the Unicode hack (auto extend 0xFFFF to 0x10FFFF) dependent on a parameter, so we only use this hack when deserializing an ATN. This avoids trouble with intervals used in other contexts (like string offsets).
- Added a few operator != overloads, to fix compilation after recent changes.
- Simplified operands comparison in SemanticContext (uses the Arrays class now). Some cleanup in that class too.
- The abstract parse tree visitor now uses const& for Any references, to avoid reallocating new instances over and over again.
- The lexer counts syntax errors the same way as the parser does. So we can directly determine if there was any error by simply examining that (which avoids having to use a temporary listener).

+++ runtime/Cpp/runtime/src/Parser.cpp
@@ -216,7 +216,7 @@ void Parser::triggerExitRuleEvent() {
   }
 }
 
-int Parser::getNumberOfSyntaxErrors() {
+size_t Parser::getNumberOfSyntaxErrors() {
   return _syntaxErrors;
 }
